refactor: Name 클래스에서 안전성 검사. TC value 일부 변경.
