+++ src/main/java/racingCar/utils/ValidateUtils.java
@@ -9,7 +9,6 @@ public class ValidateUtils {
     public static final int MIN_NAME_LENGTH = 1;
     public static final int MAX_NAME_LENGTH = 5;
     public static final String ERROR_MESSAGE_CAR_NAMES_COUNT = "자동차 이름은 1개 이상이어야 합니다.";
-    public static final String ERROR_MESSAGE_CAR_NAME_LENGTH = "자동차 이름은 1자 이상 5자 이하여야 합니다.";
     public static final String ERROR_MESSAGE_CAR_NAME_NULL_OR_EMPTY = "자동차 이름은 공백이 안됩니다.";
     public static final String ERROR_MESSAGE_ROUND_COUNT_NATURAL_NUMBER = "시도횟수는 자연수이어야 합니다.";
 
@@ -27,23 +26,20 @@ public class ValidateUtils {
         }
     }
 
-    private static void validateCarName(String carName) {
-        validateCarNameLength(carName);
-        validateCarNameNotEmpty(carName);
+    public static boolean validateCarName(String carName) {
+        return validateCarNameLength(carName) && validateCarNameNotEmpty(carName);
     }
 
-    private static void validateCarNameLength(String carName) {
-        if (carName.length() >= MIN_NAME_LENGTH && carName.length() <= MAX_NAME_LENGTH) {
-            return;
+    private static boolean validateCarNameLength(String carName) {
+        try {
+            return carName.length() >= MIN_NAME_LENGTH && carName.length() <= MAX_NAME_LENGTH;
+        } catch (NullPointerException exception){
+            throw new InvalidCarNameException(ERROR_MESSAGE_CAR_NAME_NULL_OR_EMPTY);
         }
-        throw new InvalidCarNameException(ERROR_MESSAGE_CAR_NAME_LENGTH);
     }
 
-    private static void validateCarNameNotEmpty(String carName) {
-        if (!carName.isEmpty()) {
-            return;
-        }
-        throw new InvalidCarNameException(ERROR_MESSAGE_CAR_NAME_NULL_OR_EMPTY);
+    private static boolean validateCarNameNotEmpty(String carName) {
+        return !carName.isEmpty();
     }
 
     public static int validateRacingRoundCount(String input) {
